Fixes #38836 - Add clickable indicator to status icon on All Hosts page
